Introduce RAG chat interface with authentication, WebSocket communication, and dedicated UI components
