+++ react-front/src/interfaces/RagMessageInterface.tsx
@@ -1,0 +1,9 @@
+import ShortEvent from "./ShortEventInterface";
+
+export interface RagMessage {
+    id: string;
+    text: string;
+    sender: 'user' | 'ai';
+    timestamp: Date;
+    events?: ShortEvent[];
+}
